+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Varnish Indicator Browser Extension Changelog
 
+## Unreleased
+
+-   Added support for multiple headers.
+-   If multiple x-varnish-cache headers are found, the last is used.
+-   If multiple x-varnish headers are found, all will be searched for '7707991 148236' syntax.
+
 ## v1.0.0 (2020.07.15)
 
 -   First working version.